refactor(realm): extract browser bridge from js-realm-shared
