+++ packages/webapp/src/kernel/realm/realm-ws-observer.ts
@@ -2,7 +2,7 @@
  * `realm-ws-observer.ts` — the realm-side `browser.websocket` chainable
  * observer API. Extracted from `js-realm-shared.ts`; no behavior change.
  */
-import { resolveTargetId } from './js-realm-shared.js';
+import { resolveTargetId } from './realm-browser-bridge.js';
 import { type RealmRpcClient } from './realm-rpc.js';
 import type { TabHandle, WsSelector, WsSink, WsSubscriberInfo } from './realm-types.js';
